allow scripts to dynamically create factions
* use long long instead of int for faction flags

+++ game/characters/diplomacy.h
@@ -32,8 +32,8 @@ private slots:
 
 protected:
   WorldDiplomacy::Faction* faction = nullptr;
-  unsigned int enemyFlag = 0;
-  QMetaObject::Connection factionListener;
+  unsigned long long       enemyFlag = 0;
+  QMetaObject::Connection  factionListener;
 };
 
 #endif // CHARACTERDIPLOMACY_H
